requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
lxml==5.3.0
oauthlib==3.2.2
OWSLib==0.32.0
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tweepy==4.14.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
